requirements.txt
streamlit
PyPDF2
pymupdf
easyocr
transformers
torch
scikit-learn
pandas
numpy
langchain
python-docx
openpyxl
xlrd